refac [CS] move some forms into separate dir
