+++ app/uk/gov/hmrc/agentsexternalstubsfrontend/forms/SignInRequestForm.scala
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.agentsexternalstubsfrontend.forms
+
+import play.api.data.Form
+import play.api.data.Forms.{default, ignored, mapping, nonEmptyText, optional, text}
+import play.api.libs.json.{Json, Writes}
+import uk.gov.hmrc.agentsexternalstubsfrontend.models.AuthProvider
+
+case class SignInRequest(
+  userId: String,
+  planetId: String,
+  plainTextPassword: String = "p@ssw0rd",
+  providerType: String = AuthProvider.GovernmentGateway,
+  syncToAuthLoginApi: Boolean = true
+)
+
+object SignInRequest {
+  implicit val writes: Writes[SignInRequest] = Json.writes[SignInRequest]
+}
+
+object SignInRequestForm {
+
+  val form: Form[SignInRequest] = Form[SignInRequest](
+    mapping(
+      "userId"             -> nonEmptyText,
+      "planetId"           -> nonEmptyText,
+      "password"           -> default(text, "p@ssw0rd"),
+      "providerType"       -> optional(nonEmptyText).transform[String](_.getOrElse(AuthProvider.GovernmentGateway), Some(_)),
+      "syncToAuthLoginApi" -> ignored(true)
+    )(SignInRequest.apply)(SignInRequest.unapply)
+  )
+}
