+++ it/uk/gov/hmrc/agentsexternalstubsfrontend/support/BaseISpec.scala
@@ -1,12 +1,16 @@
 package uk.gov.hmrc.agentsexternalstubsfrontend.support
 
+import akka.stream.Materializer
+import org.scalatest.OptionValues
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpecLike
 import org.scalatestplus.play.guice.GuiceOneAppPerSuite
 import play.api.Application
 import play.api.i18n.{Lang, Messages, MessagesApi}
 import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.mvc.Result
-import play.api.test.Helpers._
-import play.api.test.{DefaultAwaitTimeout, FakeRequest}
+import play.api.test.{DefaultAwaitTimeout, FakeRequest, Helpers}
 import play.twirl.api.HtmlFormat
 import uk.gov.hmrc.agentsexternalstubsfrontend.config.FrontendConfig
 import uk.gov.hmrc.agentsexternalstubsfrontend.stubs.{AuthStubs, DataStreamStubs}
@@ -14,9 +18,11 @@ import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.http.HeaderCarrierConverter
 import uk.gov.hmrc.http.HttpClient
 
+import scala.concurrent.Future
+
 class BaseISpec
-    extends UnitSpec with GuiceOneAppPerSuite with WireMockSupport with AuthStubs with DataStreamStubs
-    with MetricsTestSupport with DefaultAwaitTimeout {
+    extends AnyWordSpecLike with GuiceOneAppPerSuite with WireMockSupport with AuthStubs with DataStreamStubs
+    with MetricsTestSupport with DefaultAwaitTimeout with Matchers with OptionValues with ScalaFutures {
 
   override implicit lazy val app: Application = appBuilder.build()
 
@@ -40,7 +46,7 @@ class BaseISpec
     givenAuditConnector()
   }
 
-  protected implicit val materializer = app.materializer
+  protected implicit val materializer: Materializer = app.materializer
 
   protected def checkHtmlResultWithBodyText(result: Result, expectedSubstring: String): Unit = {
     status(result) shouldBe 200
@@ -56,5 +62,21 @@ class BaseISpec
 
   implicit def hc(implicit request: FakeRequest[_]): HeaderCarrier =
     HeaderCarrierConverter.fromRequestAndSession(request, request.session)
+
+  // the following is a collection of useful methods that should minimise
+  // the changes required when migrating away from hmrctest, which is now deprecated.
+  def status(result: Result): Int = result.header.status
+  def status(result: Future[Result]): Int = Helpers.status(result)
+  def bodyOf(result: Result): String = Helpers.contentAsString(Future.successful(result))
+  def redirectLocation(result: Result): Option[String] = Helpers.redirectLocation(Future.successful(result))
+
+  def contentType(result: Result): Option[String] =
+    result.body.contentType.map(_.split(";").take(1).mkString.trim)
+
+  def charset(result: Result): Option[String] =
+    result.body.contentType match {
+      case Some(s) if s.contains("charset=") => Some(s.split("; *charset=").drop(1).mkString.trim)
+      case _                                 => None
+    }
 
 }
