+++ test/uk/gov/hmrc/agentsexternalstubsfrontend/support/UnitSpec.scala
@@ -20,19 +20,20 @@ import org.scalatest.OptionValues
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
+import org.scalatestplus.mockito.MockitoSugar
 import play.api.mvc.Result
 import play.api.test.Helpers
 import play.api.test.Helpers.defaultAwaitTimeout
 
 import scala.concurrent.Future
 
-trait UnitSpec extends AnyWordSpecLike with Matchers with OptionValues with ScalaFutures {
+trait UnitSpec extends AnyWordSpecLike with Matchers with OptionValues with ScalaFutures with MockitoSugar {
   // the following is a collection of useful methods that should minimise
   // the changes required when migrating away from hmrctest, which is now deprecated.
   def status(result: Result): Int = result.header.status
   def status(result: Future[Result]): Int = Helpers.status(result)
   def bodyOf(result: Result): String = Helpers.contentAsString(Future.successful(result))
-  def redirectLocation(result: Result) = Helpers.redirectLocation(Future.successful(result))
+  def redirectLocation(result: Result): Option[String] = Helpers.redirectLocation(Future.successful(result))
 
   def contentType(result: Result): Option[String] =
     result.body.contentType.map(_.split(";").take(1).mkString.trim)
